Lump temporary components to one file
